requirements.txt
requests[socks]
boto3
python-dotenv
